requirements.txt
aabb==0.0.0
affine==2.4.0
affine6p==1.0.0
aiohttp==3.9.3
aiosignal==1.3.1
Ale @ file:///panfs/ccds02/nobackup/people/sberton2/ale
ansiwrap==0.8.4
anyio==4.3.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
bokeh==3.4.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.0.0
colorcet==3.0.1
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.8
dask==2024.3.1
dask-expr==1.0.5
datashader==0.16.0
datashape==0.5.2
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distributed==2024.3.1
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.15.1
fiona==1.9.5
fonttools==4.49.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.2.0
geocube==0.5.0
geopandas==0.14.3
-e git+file:///home/sberton2/git/toporing.git@d42795cb148c7c10dc3499a429aeba9cf8335dd5#egg=geotools
gmsh_interop==2021.1.1
h5py==3.10.0
idna==3.6
importlib-metadata==7.0.1
ipykernel==6.29.2
ipython==8.21.0
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json-numpy==2.0.0
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.13.0
jupyter_server_proxy==4.1.2
jupyter_server_terminals==0.5.3
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
kornia==0.7.2
kornia_rs==0.1.3
lightglue @ file:///panfs/ccds02/nobackup/people/sberton2/LightGlue
llvmlite==0.42.0
locket==1.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
meshio==5.3.5
MeshPy==2022.1.3
mistune==3.0.2
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multipledispatch==1.0.0
multiprocess==0.70.16
munch==4.0.0
nbclient==0.9.0
nbconvert==7.16.3
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
odc-geo==0.4.2
opencv-python==4.10.0.82
overrides==7.7.0
p_tqdm==1.4.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
papermill==2.5.0
param==2.0.2
parso==0.8.3
partd==1.4.1
pathos==0.3.2
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pip-system-certs==4.0
platformdirs==4.2.0
pooch==1.8.0
pox==0.3.4
ppft==1.7.6.8
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pvl==1.3.2
pyarrow==15.0.0
pycparser==2.22
pyct==0.5.0
Pygments==2.17.2
pyogrio==0.8.0
pyparsing==3.1.1
pyproj==3.6.1
pyrsistent==0.20.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pytools==2023.1.1
pytz==2024.1
pyvista==0.43.3
PyYAML==6.0.1
pyzmq==25.1.2
rasterio==1.3.9
referencing==0.33.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rioxarray==0.15.1
rpds-py==0.18.0
Rtree==1.2.0
scikit-learn==1.4.1.post1
scikit-learn-extra==0.3.0
scikit-spatial==7.2.0
scipy==1.12.0
scooby==0.9.2
seaborn==0.13.2
Send2Trash==1.8.2
-e git+file:///home/sberton2/git/shadowspy.git@d1087012e721cfc51669804c790804b1052c5265#egg=shadowspy
shapely==2.0.3
simpervisor==1.0.0
six==1.16.0
sniffio==1.3.1
snuggs==1.4.7
sortedcontainers==2.4.0
soupsieve==2.5
spiceypy==6.0.0
stack-data==0.6.3
submitit==1.5.1
sympy==1.12.1
tblib==3.0.0
tenacity==8.2.3
terminado==0.18.1
textwrap3==0.9.2
threadpoolctl==3.3.0
tinycss2==1.2.1
toolz==0.12.1
torch==2.3.1
torchvision==0.18.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
triton==2.3.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.9.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.0
vtk==9.3.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
wget==3.2
wrapt==1.16.0
xarray==2024.1.1
xarray-spatial==0.3.5
xmltodict==0.13.0
xyzservices==2023.10.1
yarl==1.9.4
zict==3.0.0
zipp==3.17.0
